move baseCommand
