+++ cli/src/commands/info/index.ts
@@ -5,7 +5,7 @@ import * as vi from "../../util/versioninfo.js";
 import { SDK } from "../../custom/globals.js";
 import { getHeader } from "../../custom/header.js";
 import { getGoVersion, getNPMVersion, getTinyGoVersion } from "../../util/systemVersions.js";
-import { BaseCommand } from "../baseCommand.js";
+import { BaseCommand } from "../../baseCommand.js";
 
 export default class InfoCommand extends BaseCommand {
   static args = {};
