+++ cli/src/commands/new/index.ts
@@ -25,7 +25,7 @@ import { getHeader } from "../../custom/header.js";
 import * as inquirer from "@inquirer/prompts";
 import { getGoVersion, getTinyGoVersion } from "../../util/systemVersions.js";
 import { generateAppName } from "../../util/appname.js";
-import { BaseCommand } from "../baseCommand.js";
+import { BaseCommand } from "../../baseCommand.js";
 
 const MODUS_DEFAULT_TEMPLATE_NAME = "default";
