+++ cli/src/commands/runtime/list/index.ts
@@ -11,7 +11,7 @@ import { Command, Flags } from "@oclif/core";
 import chalk from "chalk";
 import * as vi from "../../../util/versioninfo.js";
 import { getHeader } from "../../../custom/header.js";
-import { BaseCommand } from "../../baseCommand.js";
+import { BaseCommand } from "../../../baseCommand.js";
 
 export default class RuntimeListCommand extends BaseCommand {
   static args = {};
